requirements.txt
flask
numpy
scikit-learn
ifaddr
itsdangerous
Jinja2
MarkupSafe
Werkzeug
zeroconf
gunicorn
pyyaml 
pandas 
joblib

